+++ alg/line.cpp
@@ -190,7 +190,7 @@ bool LineParticle::hasTailFollower() const {
   return labelOfFirstNbrWithProperty<LineParticle>(propertyCheck) != -1;
 }
 
-LineSystem::LineSystem(int numParticles, float holeProb) {
+LineSystem::LineSystem(int numParticles, double holeProb) {
   Q_ASSERT(numParticles > 0);
   Q_ASSERT(0 <= holeProb && holeProb <= 1);
 
@@ -224,7 +224,7 @@ LineSystem::LineSystem(int numParticles, float holeProb) {
     occupied.insert(randomCandidate);
 
     // Add this candidate as a particle if not a hole.
-    if (randBool(1.0f - holeProb)) {
+    if (randBool(1.0 - holeProb)) {
       insert(new LineParticle(randomCandidate, -1, randDir(), *this,
                               LineParticle::State::Idle));
       ++numNonStaticParticles;
